requirements/test.txt
coverage>=3.7.1
django-nose>=1.3
django-dynamic-fixture>=1.8.1
flake8>=2.2.0
mock>=1.0.1
nose>=1.3.0
